requirements.txt
ruamel.yaml
python-decouple
humanfriendly
numpy